Fix problems when building with msvc on x86_64
I honestly don't know why I didn't run into the __USER_LABEL_PREFIX__ problem before. - This shouldn't be the first time it would have been a problem...

+++ src/deemon/runtime/slab-impl.c.inl
@@ -348,7 +348,7 @@ FORCELOCAL void
   ASSERTF((((uintptr_t)ptr - (uintptr_t)page) % ITEMSIZE) == 0,
           "Invalid slab-pointer %p is improperly aligned for slab of size %#Ix",
           ptr,(size_t)ITEMSIZE);
-  index = ((uintptr_t)ptr - (uintptr_t)page) / ITEMSIZE;
+  index = (unsigned int)(((uintptr_t)ptr - (uintptr_t)page) / ITEMSIZE);
   ASSERTF(index < (unsigned int)SLAB_PAGECOUNT,
           "Invalid slab-pointer %p is part of slab page controller (index = %u/%u)",
           ptr,index,(unsigned int)SLAB_PAGECOUNT);
